(description)
We have changed this axiom (roh:hasKnowledgeArea only skop:Concept) to this other  (  (skos:Concept and (skos:inScheme some roh:KnowledgeArea)) in all classes that have it

the modification of this axiom should not affect because the range of the object property roh:hasKnowledgeArea is (skos:Concept and (skos:inScheme some roh:KnowledgeArea).
We did it to be coherent.